# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Event Manager",
    "description": "Event Manager is an application where users who are managed by admin can create events and choose a center where they want the event to take place",
    "version": "1.0",
    "contact": "email: kenolusola@gmail.com"
  },
  "servers": [
    {
      "url": "https://emanager1980.herokuapp.com",
      "variables": {}
    }
  ],
  "paths": {
    "/api/v1/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get All Users",
        "description": "API endpoint to get all users in the application",
        "operationId": "GetAllUsers",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyaWQiOjEsImlzYWRtaW4iOnRydWUsInVzZXJyb2xlIjoic3VwZXJhZG1pbiIsImlhdCI6MTUyNjQwOTkxNywiZXhwIjoxNTI2NTgyNzE3fQ.vJWnidUiNieFE-tof_MiJTBWad5trZWQyrnPEY9pkbI"
            },
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyaWQiOjEsImlzYWRtaW4iOnRydWUsInVzZXJyb2xlIjoic3VwZXJhZG1pbiIsImlhdCI6MTUyNjQwOTkxNywiZXhwIjoxNTI2NTgyNzE3fQ.vJWnidUiNieFE-tof_MiJTBWad5trZWQyrnPEY9pkbI"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found all users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "status": "Success",
                    "message": "These are all user details",
                    "data": {
                      "allCenters": []
                    }
                  }
                }
              }
            } 
          },
          "401": {
            "description": "Users not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "status": "Error",
                    "message": "You are not authorised to view this information"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Users not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "status": "Error",
                    "message": "users not found"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "status": "Error",
                    "message": "Server Error"
                  }
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [],
        "x-operation-settings": {
          "CollectParameters": false,
          "AllowDynamicQueryParameters": false,
          "AllowDynamicFormParameters": false,
          "IsMultiContentStreaming": false
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Sign Up, creates new users",
        "description": "Allows users of the application to sign up",
        "operationId": "SignUp",
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "example": "application/x-www-form-urlencoded"
            },
            "example": "application/x-www-form-urlencoded"
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "required": [
                  "firstname",
                  "lastname",
                  "email",
                  "password"
                ],
                "type": "object",
                "properties": {
                  "firstname": {
                    "type": "string",
                    "example": "Bala"
                  },
                  "lastname": {
                    "type": "string",
                    "example": "Ahmed"
                  },
                  "email": {
                    "type": "string",
                    "example": "bala@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "1234567890"
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "Success",
                    "message": "User successfully signed up",
                    "data": {
                      "firstname": "kunle",
                      "lastname": "williams",
                      "role": "user",
                      "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyaWQiOjQsImlzYWRtaW4iOmZhbHNlLCJ1c2Vycm9sZSI6InVzZXIiLCJpYXQiOjE1MjY0MDgzMzMsImV4cCI6MTUyNjU4MTEzM30.ziozx4c9yxvPDeGBefawbHEzeB0dh6Kz1-2EkNyvb-4"
                    }
                  },
                  "$ref": "#/components/schemas/SignUp"
                },
                "example": {
                  "status": "Success",
                  "message": "User successfully signed up",
                  "data": {
                    "firstname": "kunle",
                    "lastname": "williams",
                    "role": "user",
                    "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyaWQiOjQsImlzYWRtaW4iOmZhbHNlLCJ1c2Vycm9sZSI6InVzZXIiLCJpYXQiOjE1MjY0MDgzMzMsImV4cCI6MTUyNjU4MTEzM30.ziozx4c9yxvPDeGBefawbHEzeB0dh6Kz1-2EkNyvb-4"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "status": "Error",
                    "message": "Server Error"
                  }
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [],
        "x-operation-settings": {
          "CollectParameters": false,
          "AllowDynamicQueryParameters": false,
          "AllowDynamicFormParameters": false,
          "IsMultiContentStreaming": false
        }
      }
    },
...
  }
}